Refactor parallel classes to use shared mixins
